ElementTraversal/tests/submissions/W3C/Element-lastElementChild.xhtml

<!DOCTYPE html>
<html lang="en" xml:lang="en" xmlns='http://www.w3.org/1999/xhtml' xmlns:pickle="http://ns.example.org/pickle" >
<head>
	<meta charset="utf-8"/>

	<title id='title'>lastElementChild</title>
	<script src="/resources/testharnessreport.js"></script>	
	<script src="/resources/testharness.js"></script>
</head>

<body>

  <h1 id='h1'>Test of lastElementChild</h1>
  <p id="parentEl">This <span id="first_element_child"><span>is</span> a <span>test</span></span> for the 
  <span id="middle_element_child">ElementTraversal specification</span>  
  <span id="last_element_child" style="display:none;">display: none &#160;</span>.</p>

  <div id="divEl">
    <pickle:dill />
  </div>


  <div id="log">Running test...</div>

	<script>
	  test(function() {
                assert_true(!!document.getElementById("parentEl").lastElementChild);
               }, "lastElementChild must exist");
	  test(function() {
                assert_equals( document.getElementById("title").lastElementChild, null);
                assert_equals( document.getElementById("h1").lastElementChild, null);
               }, "lastElementChild must return null if no child element");
	  test(function() {
	        var parentEl = document.getElementById("parentEl");	        
                assert_not_equals( parentEl.lastElementChild, null);
                assert_equals( parentEl.lastElementChild.getAttribute("id"), "last_element_child");
	        var divEl = document.getElementById("divEl");
                assert_not_equals( divEl.lastElementChild, null);
                assert_equals( divEl.lastElementChild.prefix, "pickle");
               }, "lastElementChild must return the last element child");
	  test(function() {
	        var parentEl = document.getElementById("parentEl");	        
                assert_equals( parentEl.lastElementChild.getAttribute("id"), "last_element_child");
	        var span = document.createElement("span");
	        span.setAttribute("id", "dummy");
	        parentEl.appendChild(span);
                assert_not_equals(parentEl.lastElementChild, null);
                assert_equals( parentEl.lastElementChild.getAttribute("id"), "dummy");
	        parentEl.removeChild(span);
                assert_equals( parentEl.lastElementChild.getAttribute("id"), "last_element_child");
               }, "lastElementChild must be dynamically updated");
	  test(function() {
                assert_readonly(document.getElementById("parentEl"), "lastElementChild");
               }, "lastElementChild must be readonly");
	</script>
</body>
</html>
